First version
we have a list of cats and whenever we click on one, the number of
clicks goes up and saved for the selected cat. Also afterwards, some
Refactoring was done in a way that it would easily be scalable.
separating View and Model. connecting them through an interface called
Octopus.